Split tests for consumer and processor
It is better to split tests because we are coupling the way we processs
the events with the technology to receive them.

It is better to test in isolation (these are not integration tests) and
then split the different scenarios (create/update/delete)
